// src/app/core/consts/report-response-type.enum.ts
export enum ReportResponseType {
  ERROR_REPORT = 'ERROR_REPORT',
  ERROR_COMPETENCY_REPORT = 'ERROR_COMPETENCY_REPORT',
  ERROR_JOB_REPORT = 'ERROR_JOB_REPORT',
  ERROR_AREA_INDIVIDUAL_REPORT = 'ERROR_AREA_INDIVIDUAL_REPORT',
  ERROR_GROUPING_INDIVIDUAL_REPORT = 'ERROR_GROUPING_INDIVIDUAL_REPORT',

  ERROR_UPDATE_REPORT = 'ERROR_UPDATE_REPORT',
  ERROR_UPDATE_COMPETENCY_REPORT = 'ERROR_UPDATE_COMPETENCY_REPORT',
  ERROR_UPDATE_JOB_REPORT = 'ERROR_UPDATE_JOB_REPORT',
  ERROR_UPDATE_AREA_INDIVIDUAL_REPORT = 'ERROR_UPDATE_AREA_INDIVIDUAL_REPORT',
  ERROR_UPDATE_GROUPING_INDIVIDUAL_REPORT = 'ERROR_UPDATE_GROUPING_INDIVIDUAL_REPORT',

  ERROR = 'ERROR',
  SUCCESS = 'SUCCESS',
  WARNING = 'WARNING',
}